requirements.txt
streamlit>=1.50.0
google-genai>=1.0.0
supabase>=2.27.0
pinecone>=7.0.0
neo4j>=5.25.0
minio>=7.2.20
reportlab>=4.4.0
weasyprint>=66.0
PyMuPDF>=1.26.0
easyocr>=1.7.0
pillow>=11.0.0
jinja2>=3.1.0
pydantic>=2.9.0
pydantic-settings>=2.11.0
python-dotenv>=1.1.0
pandas>=2.2.0
matplotlib>=3.9.0
